+++ service/controller/resource/monitoring/scrapeconfigs/resource.go
@@ -30,8 +30,6 @@ const (
 	unknownObservabilityBundleVersion = "0.0.0"
 )
 
-var kubernetesTargets = []string{"kube-apiserver", "kube-controller-manager", "kube-scheduler"}
-
 type Config struct {
 	K8sClient k8sclient.Interface
 	Logger    micrologger.Logger
@@ -208,10 +206,14 @@ func getObservabilityBundleAppVersion(ctx context.Context, ctrlClient client.Cli
 	appName := fmt.Sprintf("%s-observability-bundle", key.ClusterID(cluster))
 	appNamespace := cluster.GetNamespace()
 
-	if key.IsManagementCluster(config.Installation, cluster) && !key.IsCAPIManagementCluster(config.Provider) {
-		// Vintage MC
-		appName = "observability-bundle"
-		appNamespace = "giantswarm"
+	if !key.IsCAPIManagementCluster(config.Provider) {
+		if key.IsManagementCluster(config.Installation, cluster) {
+			// Vintage MC
+			appName = "observability-bundle"
+			appNamespace = "giantswarm"
+		} else {
+			appNamespace = key.ClusterID(cluster)
+		}
 	}
 
 	app := &appsv1alpha1.App{}
@@ -256,11 +258,11 @@ func listTargetsToIgnore(ctx context.Context, ctrlClient client.Client, cluster 
 	}
 
 	if version.GTE(initialBundleVersion) {
-		ignoredTargets = append(ignoredTargets, kubernetesTargets...)
+		ignoredTargets = append(ignoredTargets, "prometheus-operator-app", "kube-apiserver", "kube-controller-manager", "kube-scheduler")
 	}
 
 	if version.GTE(bundleWithKSMAndExportersVersion) {
-		ignoredTargets = append(ignoredTargets, "kubelet", "coredns", "kube-state-metrics")
+		ignoredTargets = append(ignoredTargets, "kubelet", "coredns", "kube-state-metrics", "aws-load-balancer-controller")
 
 		if key.IsCAPIManagementCluster(config.Provider) {
 			ignoredTargets = append(ignoredTargets, "etcd")